[REF] moved filepath related function to utils class
