+++ src/odoo_configurator/utils.py
@@ -3,7 +3,7 @@
 
 from .logging import get_logger
 import os
-
+import sys
 
 class Utils:
     """This class is intended to provide utility functions which are not related to a specific configurator app.
@@ -19,3 +19,29 @@ class Utils:
         if not var:
             self.logger.warning(f"{var} Environment Variable not found or empty")
         return var
+
+    @staticmethod
+    def get_file_full_path(path):
+        if not path:
+            return ''
+        param_path = path
+        if not os.path.isfile(path):
+            path = os.path.join(os.path.dirname(sys.argv[1]), param_path)
+        if not os.path.isfile(path):
+            path = os.path.join(os.path.dirname(sys.argv[1]), 'datas', param_path)
+        if not os.path.isfile(path):
+            raise FileNotFoundError('%s not found!' % param_path)
+        return path
+
+    @staticmethod
+    def get_dir_full_path(path):
+        if not path:
+            return ''
+        param_path = path
+        if not os.path.isdir(path):
+            path = os.path.join(os.path.dirname(sys.argv[1]), param_path)
+        if not os.path.isdir(path):
+            path = os.path.join(os.path.dirname(sys.argv[1]), 'datas', param_path)
+        if not os.path.isdir(path):
+            raise NotADirectoryError('%s not found!' % param_path)
+        return path
